preparing code extraction for modularity & annetto instantation
